requirements.txt
openai
SpeechRecognition
pyttsx3
PyAudio
PyYAML